refactor: move routes to separate file
